Moved the resource reloading query string logic into a new type.
